\encoding{latin1}
\name{CHorANADUP}
\alias{CHorANADUP}
\docType{data}
\title{Analytical duplicates of the C-horizon Kola data}
\description{
Analytical duplicates have been selected for quality control.
}
\usage{data(CHorANADUP)}
\format{
  A data frame with 52 observations on the following 190 variables.
  \describe{
    \item{\code{A1_.Loc}}{a numeric vector}
    \item{\code{A2_.Loc}}{a numeric vector}
    \item{\code{A1_Ag}}{a numeric vector}
    \item{\code{A1_Ag_INAA}}{a numeric vector}
    \item{\code{A1_Al}}{a numeric vector}
    \item{\code{A1_Al2O3}}{a numeric vector}
    \item{\code{A1_As}}{a numeric vector}
    \item{\code{A1_As_INAA}}{a numeric vector}
    \item{\code{A1_Au_INAA}}{a numeric vector}
    \item{\code{A1_B}}{a numeric vector}
    \item{\code{A1_Ba}}{a numeric vector}
    \item{\code{A1_Ba_INAA}}{a numeric vector}
    \item{\code{A1_Be}}{a numeric vector}
    \item{\code{A1_Bi}}{a numeric vector}
    \item{\code{A1_Br}}{a numeric vector}
    \item{\code{A1_Br_INAA}}{a numeric vector}
    \item{\code{A1_Ca}}{a numeric vector}
    \item{\code{A1_Ca_INAA}}{a numeric vector}
    \item{\code{A1_CaO}}{a numeric vector}
    \item{\code{A1_Cd}}{a numeric vector}
    \item{\code{A1_Ce_INAA}}{a numeric vector}
    \item{\code{A1_Cl}}{a numeric vector}
    \item{\code{A1_Co}}{a numeric vector}
    \item{\code{A1_Co_INAA}}{a numeric vector}
    \item{\code{A1_Cond}}{a numeric vector}
    \item{\code{A1_Cr}}{a numeric vector}
    \item{\code{A1_Cr_INAA}}{a numeric vector}
    \item{\code{A1_Cs_INAA}}{a numeric vector}
    \item{\code{A1_Cu}}{a numeric vector}
    \item{\code{A1_Eu_INAA}}{a numeric vector}
    \item{\code{A1_F}}{a numeric vector}
    \item{\code{A1_F_ionselect}}{a numeric vector}
    \item{\code{A1_Fe}}{a numeric vector}
    \item{\code{A1_Fe_INAA}}{a numeric vector}
    \item{\code{A1_Fe2O3}}{a numeric vector}
    \item{\code{A1_Hf_INAA}}{a numeric vector}
    \item{\code{A1_Hg}}{a numeric vector}
    \item{\code{A1_Hg_INAA}}{a numeric vector}
    \item{\code{A1_Ir_INAA}}{a numeric vector}
    \item{\code{A1_K}}{a numeric vector}
    \item{\code{A1_K2O}}{a numeric vector}
    \item{\code{A1_La}}{a numeric vector}
    \item{\code{A1_La_INAA}}{a numeric vector}
    \item{\code{A1_Li}}{a numeric vector}
    \item{\code{A1_LOI}}{a numeric vector}
    \item{\code{A1_Lu_INAA}}{a numeric vector}
    \item{\code{A1_Mass_INAA}}{a numeric vector}
    \item{\code{A1_Mg}}{a numeric vector}
    \item{\code{A1_MgO}}{a numeric vector}
    \item{\code{A1_Mn}}{a numeric vector}
    \item{\code{A1_MnO}}{a numeric vector}
    \item{\code{A1_Mo}}{a numeric vector}
    \item{\code{A1_Mo_INAA}}{a numeric vector}
    \item{\code{A1_Na}}{a numeric vector}
    \item{\code{A1_Na_INAA}}{a numeric vector}
    \item{\code{A1_Na2O}}{a numeric vector}
    \item{\code{A1_Nd_INAA}}{a numeric vector}
    \item{\code{A1_Ni}}{a numeric vector}
    \item{\code{A1_Ni_INAA}}{a numeric vector}
    \item{\code{A1_NO2}}{a numeric vector}
    \item{\code{A1_NO3}}{a numeric vector}
    \item{\code{A1_P}}{a numeric vector}
    \item{\code{A1_P2O5}}{a numeric vector}
    \item{\code{A1_Pb}}{a numeric vector}
    \item{\code{A1_pH}}{a numeric vector}
    \item{\code{A1_PO4}}{a numeric vector}
    \item{\code{A1_Rb}}{a numeric vector}
    \item{\code{A1_S}}{a numeric vector}
    \item{\code{A1_Sb}}{a numeric vector}
    \item{\code{A1_Sb_INAA}}{a numeric vector}
    \item{\code{A1_Sc}}{a numeric vector}
    \item{\code{A1_Sc_INAA}}{a numeric vector}
    \item{\code{A1_Se}}{a numeric vector}
    \item{\code{A1_Se_INAA}}{a numeric vector}
    \item{\code{A1_Si}}{a numeric vector}
    \item{\code{A1_SiO2}}{a numeric vector}
    \item{\code{A1_Sm_INAA}}{a numeric vector}
    \item{\code{A1_Sn_INAA}}{a numeric vector}
    \item{\code{A1_SO4}}{a numeric vector}
    \item{\code{A1_Sr}}{a numeric vector}
    \item{\code{A1_Sr_INAA}}{a numeric vector}
    \item{\code{A1_Sum}}{a numeric vector}
    \item{\code{A1_Ta_INAA}}{a numeric vector}
    \item{\code{A1_Tb_INAA}}{a numeric vector}
    \item{\code{A1_Te}}{a numeric vector}
    \item{\code{A1_Th}}{a numeric vector}
    \item{\code{A1_Th_INAA}}{a numeric vector}
    \item{\code{A1_Ti}}{a numeric vector}
    \item{\code{A1_TiO2}}{a numeric vector}
    \item{\code{A1_U_INAA}}{a numeric vector}
    \item{\code{A1_V}}{a numeric vector}
    \item{\code{A1_W_INAA}}{a numeric vector}
    \item{\code{A1_Y}}{a numeric vector}
    \item{\code{A1_Yb_INAA}}{a numeric vector}
    \item{\code{A1_Zn}}{a numeric vector}
    \item{\code{A1_Zn_INAA}}{a numeric vector}
    \item{\code{A2_Ag}}{a numeric vector}
    \item{\code{A2_Ag_INAA}}{a numeric vector}
    \item{\code{A2_Al}}{a numeric vector}
    \item{\code{A2_Al2O3}}{a numeric vector}
    \item{\code{A2_As}}{a numeric vector}
    \item{\code{A2_As_INAA}}{a numeric vector}
    \item{\code{A2_Au_INAA}}{a numeric vector}
    \item{\code{A2_B}}{a numeric vector}
    \item{\code{A2_Ba}}{a numeric vector}
    \item{\code{A2_Ba_INAA}}{a numeric vector}
    \item{\code{A2_Be}}{a numeric vector}
    \item{\code{A2_Bi}}{a numeric vector}
    \item{\code{A2_Br}}{a numeric vector}
    \item{\code{A2_Br_INAA}}{a numeric vector}
    \item{\code{A2_Ca}}{a numeric vector}
    \item{\code{A2_Ca_INAA}}{a numeric vector}
    \item{\code{A2_CaO}}{a numeric vector}
    \item{\code{A2_Cd}}{a numeric vector}
    \item{\code{A2_Ce_INAA}}{a numeric vector}
    \item{\code{A2_Cl}}{a numeric vector}
    \item{\code{A2_Co}}{a numeric vector}
    \item{\code{A2_Co_INAA}}{a numeric vector}
    \item{\code{A2_Cond}}{a numeric vector}
    \item{\code{A2_Cr}}{a numeric vector}
    \item{\code{A2_Cr_INAA}}{a numeric vector}
    \item{\code{A2_Cs_INAA}}{a numeric vector}
    \item{\code{A2_Cu}}{a numeric vector}
    \item{\code{A2_Eu_INAA}}{a numeric vector}
    \item{\code{A2_F}}{a numeric vector}
    \item{\code{A2_F_ionselect}}{a numeric vector}
    \item{\code{A2_Fe}}{a numeric vector}
    \item{\code{A2_Fe_INAA}}{a numeric vector}
    \item{\code{A2_Fe2O3}}{a numeric vector}
    \item{\code{A2_Hf_INAA}}{a numeric vector}
    \item{\code{A2_Hg}}{a numeric vector}
    \item{\code{A2_Hg_INAA}}{a numeric vector}
    \item{\code{A2_Ir_INAA}}{a numeric vector}
    \item{\code{A2_K}}{a numeric vector}
    \item{\code{A2_K2O}}{a numeric vector}
    \item{\code{A2_La}}{a numeric vector}
    \item{\code{A2_La_INAA}}{a numeric vector}
    \item{\code{A2_Li}}{a numeric vector}
    \item{\code{A2_LOI}}{a numeric vector}
    \item{\code{A2_Lu_INAA}}{a numeric vector}
    \item{\code{A2_Mass_INAA}}{a numeric vector}
    \item{\code{A2_Mg}}{a numeric vector}
    \item{\code{A2_MgO}}{a numeric vector}
    \item{\code{A2_Mn}}{a numeric vector}
    \item{\code{A2_MnO}}{a numeric vector}
    \item{\code{A2_Mo}}{a numeric vector}
    \item{\code{A2_Mo_INAA}}{a numeric vector}
    \item{\code{A2_Na}}{a numeric vector}
    \item{\code{A2_Na_INAA}}{a numeric vector}
    \item{\code{A2_Na2O}}{a numeric vector}
    \item{\code{A2_Nd_INAA}}{a numeric vector}
    \item{\code{A2_Ni}}{a numeric vector}
    \item{\code{A2_Ni_INAA}}{a numeric vector}
    \item{\code{A2_NO2}}{a numeric vector}
    \item{\code{A2_NO3}}{a numeric vector}
    \item{\code{A2_P}}{a numeric vector}
    \item{\code{A2_P2O5}}{a numeric vector}
    \item{\code{A2_Pb}}{a numeric vector}
    \item{\code{A2_pH}}{a numeric vector}
    \item{\code{A2_PO4}}{a numeric vector}
    \item{\code{A2_Rb}}{a numeric vector}
    \item{\code{A2_S}}{a numeric vector}
    \item{\code{A2_Sb}}{a numeric vector}
    \item{\code{A2_Sb_INAA}}{a numeric vector}
    \item{\code{A2_Sc}}{a numeric vector}
    \item{\code{A2_Sc_INAA}}{a numeric vector}
    \item{\code{A2_Se}}{a numeric vector}
    \item{\code{A2_Se_INAA}}{a numeric vector}
    \item{\code{A2_Si}}{a numeric vector}
    \item{\code{A2_SiO2}}{a numeric vector}
    \item{\code{A2_Sm_INAA}}{a numeric vector}
    \item{\code{A2_Sn_INAA}}{a numeric vector}
    \item{\code{A2_SO4}}{a numeric vector}
    \item{\code{A2_Sr}}{a numeric vector}
    \item{\code{A2_Sr_INAA}}{a numeric vector}
    \item{\code{A2_Sum}}{a numeric vector}
    \item{\code{A2_Ta_INAA}}{a numeric vector}
    \item{\code{A2_Tb_INAA}}{a numeric vector}
    \item{\code{A2_Te}}{a numeric vector}
    \item{\code{A2_Th}}{a numeric vector}
    \item{\code{A2_Th_INAA}}{a numeric vector}
    \item{\code{A2_Ti}}{a numeric vector}
    \item{\code{A2_TiO2}}{a numeric vector}
    \item{\code{A2_U_INAA}}{a numeric vector}
    \item{\code{A2_V}}{a numeric vector}
    \item{\code{A2_W_INAA}}{a numeric vector}
    \item{\code{A2_Y}}{a numeric vector}
    \item{\code{A2_Yb_INAA}}{a numeric vector}
    \item{\code{A2_Zn}}{a numeric vector}
    \item{\code{A2_Zn_INAA}}{a numeric vector}
  }
}
\source{
Kola Project (1993-1998)
}
\references{
Reimann C, Äyräs M, Chekushin V, Bogatyrev I, Boyd R, Caritat P de, Dutter R,
Finne TE, Halleraker JH, Jæger Ø, Kashulina G, Lehto O, Niskavaara H, Pavlov V,
Räisänen ML, Strand T, Volden T. Environmental Geochemical Atlas of the Central
Barents Region. NGU-GTK-CKE Special Publication, Geological Survey of Norway,
Trondheim, Norway, 1998.
}
\examples{
data(CHorANADUP)
str(CHorANADUP)
}
\keyword{datasets}
