\encoding{latin1}
\name{CHorFieldDUP}
\alias{CHorFieldDUP}
\docType{data}
\title{Field duplicates of the C-horizon Kola data}
\description{
Field duplicates have been selected for quality control.
}
\usage{data(CHorFieldDUP)}
\format{
  A data frame with 49 observations on the following 240 variables.
  \describe{
    \item{\code{F1_.Loc}}{a numeric vector}
    \item{\code{F2_.Loc}}{a numeric vector}
    \item{\code{XCOO}}{a numeric vector}
    \item{\code{YCOO}}{a numeric vector}
    \item{\code{F1_Ag}}{a numeric vector}
    \item{\code{F1_Ag_INAA}}{a numeric vector}
    \item{\code{F1_Al}}{a numeric vector}
    \item{\code{F1_Al2O3}}{a numeric vector}
    \item{\code{F1_As}}{a numeric vector}
    \item{\code{F1_As_INAA}}{a numeric vector}
    \item{\code{F1_Au_INAA}}{a numeric vector}
    \item{\code{F1_B}}{a numeric vector}
    \item{\code{F1_Ba}}{a numeric vector}
    \item{\code{F1_Ba_INAA}}{a numeric vector}
    \item{\code{F1_Be}}{a numeric vector}
    \item{\code{F1_Bi}}{a numeric vector}
    \item{\code{F1_Br}}{a numeric vector}
    \item{\code{F1_Br_INAA}}{a numeric vector}
    \item{\code{F1_Ca}}{a numeric vector}
    \item{\code{F1_Ca_INAA}}{a numeric vector}
    \item{\code{F1_CaO}}{a numeric vector}
    \item{\code{F1_Cd}}{a numeric vector}
    \item{\code{F1_Ce_INAA}}{a numeric vector}
    \item{\code{F1_Cl}}{a numeric vector}
    \item{\code{F1_Co}}{a numeric vector}
    \item{\code{F1_Co_INAA}}{a numeric vector}
    \item{\code{F1_Cond}}{a numeric vector}
    \item{\code{F1_Cr}}{a numeric vector}
    \item{\code{F1_Cr_INAA}}{a numeric vector}
    \item{\code{F1_Cs_INAA}}{a numeric vector}
    \item{\code{F1_Cu}}{a numeric vector}
    \item{\code{F1_Eu_INAA}}{a numeric vector}
    \item{\code{F1_F}}{a numeric vector}
    \item{\code{F1_F_ionselect}}{a numeric vector}
    \item{\code{F1_Fe}}{a numeric vector}
    \item{\code{F1_Fe_INAA}}{a numeric vector}
    \item{\code{F1_Fe2O3}}{a numeric vector}
    \item{\code{F1_Hf_INAA}}{a numeric vector}
    \item{\code{F1_Hg}}{a numeric vector}
    \item{\code{F1_Hg_INAA}}{a numeric vector}
    \item{\code{F1_Ir_INAA}}{a numeric vector}
    \item{\code{F1_K}}{a numeric vector}
    \item{\code{F1_K2O}}{a numeric vector}
    \item{\code{F1_La}}{a numeric vector}
    \item{\code{F1_La_INAA}}{a numeric vector}
    \item{\code{F1_Li}}{a numeric vector}
    \item{\code{F1_LOI}}{a numeric vector}
    \item{\code{F1_Lu_INAA}}{a numeric vector}
    \item{\code{F1_Mass_INAA}}{a numeric vector}
    \item{\code{F1_Mg}}{a numeric vector}
    \item{\code{F1_MgO}}{a numeric vector}
    \item{\code{F1_Mn}}{a numeric vector}
    \item{\code{F1_MnO}}{a numeric vector}
    \item{\code{F1_Mo}}{a numeric vector}
    \item{\code{F1_Mo_INAA}}{a numeric vector}
    \item{\code{F1_Na}}{a numeric vector}
    \item{\code{F1_Na_INAA}}{a numeric vector}
    \item{\code{F1_Na2O}}{a numeric vector}
    \item{\code{F1_Nd_INAA}}{a numeric vector}
    \item{\code{F1_Ni}}{a numeric vector}
    \item{\code{F1_Ni_INAA}}{a numeric vector}
    \item{\code{F1_NO2}}{a numeric vector}
    \item{\code{F1_NO3}}{a numeric vector}
    \item{\code{F1_P}}{a numeric vector}
    \item{\code{F1_P2O5}}{a numeric vector}
    \item{\code{F1_Pb}}{a numeric vector}
    \item{\code{F1_pH}}{a numeric vector}
    \item{\code{F1_PO4}}{a numeric vector}
    \item{\code{F1_Rb}}{a numeric vector}
    \item{\code{F1_S}}{a numeric vector}
    \item{\code{F1_Sb}}{a numeric vector}
    \item{\code{F1_Sb_INAA}}{a numeric vector}
    \item{\code{F1_Sc}}{a numeric vector}
    \item{\code{F1_Sc_INAA}}{a numeric vector}
    \item{\code{F1_Se}}{a numeric vector}
    \item{\code{F1_Se_INAA}}{a numeric vector}
    \item{\code{F1_Si}}{a numeric vector}
    \item{\code{F1_SiO2}}{a numeric vector}
    \item{\code{F1_Sm_INAA}}{a numeric vector}
    \item{\code{F1_Sn_INAA}}{a numeric vector}
    \item{\code{F1_SO4}}{a numeric vector}
    \item{\code{F1_Sr}}{a numeric vector}
    \item{\code{F1_Sr_INAA}}{a numeric vector}
    \item{\code{F1_Sum}}{a numeric vector}
    \item{\code{F1_Ta_INAA}}{a numeric vector}
    \item{\code{F1_Tb_INAA}}{a numeric vector}
    \item{\code{F1_Te}}{a numeric vector}
    \item{\code{F1_Th}}{a numeric vector}
    \item{\code{F1_Th_INAA}}{a numeric vector}
    \item{\code{F1_Ti}}{a numeric vector}
    \item{\code{F1_TiO2}}{a numeric vector}
    \item{\code{F1_U_INAA}}{a numeric vector}
    \item{\code{F1_V}}{a numeric vector}
    \item{\code{F1_W_INAA}}{a numeric vector}
    \item{\code{F1_Y}}{a numeric vector}
    \item{\code{F1_Yb_INAA}}{a numeric vector}
    \item{\code{F1_Zn}}{a numeric vector}
    \item{\code{F1_Zn_INAA}}{a numeric vector}
    \item{\code{F2_Ag}}{a numeric vector}
    \item{\code{F2_Ag_INAA}}{a numeric vector}
    \item{\code{F2_Al}}{a numeric vector}
    \item{\code{F2_Al2O3}}{a numeric vector}
    \item{\code{F2_As}}{a numeric vector}
    \item{\code{F2_As_INAA}}{a numeric vector}
    \item{\code{F2_Au_INAA}}{a numeric vector}
    \item{\code{F2_B}}{a numeric vector}
    \item{\code{F2_Ba}}{a numeric vector}
    \item{\code{F2_Ba_INAA}}{a numeric vector}
    \item{\code{F2_Be}}{a numeric vector}
    \item{\code{F2_Bi}}{a numeric vector}
    \item{\code{F2_Br}}{a numeric vector}
    \item{\code{F2_Br_INAA}}{a numeric vector}
    \item{\code{F2_Ca}}{a numeric vector}
    \item{\code{F2_Ca_INAA}}{a numeric vector}
    \item{\code{F2_CaO}}{a numeric vector}
    \item{\code{F2_Cd}}{a numeric vector}
    \item{\code{F2_Ce_INAA}}{a numeric vector}
    \item{\code{F2_Cl}}{a numeric vector}
    \item{\code{F2_Co}}{a numeric vector}
    \item{\code{F2_Co_INAA}}{a numeric vector}
    \item{\code{F2_Cond}}{a numeric vector}
    \item{\code{F2_Cr}}{a numeric vector}
    \item{\code{F2_Cr_INAA}}{a numeric vector}
    \item{\code{F2_Cs_INAA}}{a numeric vector}
    \item{\code{F2_Cu}}{a numeric vector}
    \item{\code{F2_Eu_INAA}}{a numeric vector}
    \item{\code{F2_F}}{a numeric vector}
    \item{\code{F2_F_ionselect}}{a numeric vector}
    \item{\code{F2_Fe}}{a numeric vector}
    \item{\code{F2_Fe_INAA}}{a numeric vector}
    \item{\code{F2_Fe2O3}}{a numeric vector}
    \item{\code{F2_Hf_INAA}}{a numeric vector}
    \item{\code{F2_Hg}}{a numeric vector}
    \item{\code{F2_Hg_INAA}}{a numeric vector}
    \item{\code{F2_Ir_INAA}}{a numeric vector}
    \item{\code{F2_K}}{a numeric vector}
    \item{\code{F2_K2O}}{a numeric vector}
    \item{\code{F2_La}}{a numeric vector}
    \item{\code{F2_La_INAA}}{a numeric vector}
    \item{\code{F2_Li}}{a numeric vector}
    \item{\code{F2_LOI}}{a numeric vector}
    \item{\code{F2_Lu_INAA}}{a numeric vector}
    \item{\code{F2_Mass_INAA}}{a numeric vector}
    \item{\code{F2_Mg}}{a numeric vector}
    \item{\code{F2_MgO}}{a numeric vector}
    \item{\code{F2_Mn}}{a numeric vector}
    \item{\code{F2_MnO}}{a numeric vector}
    \item{\code{F2_Mo}}{a numeric vector}
    \item{\code{F2_Mo_INAA}}{a numeric vector}
    \item{\code{F2_Na}}{a numeric vector}
    \item{\code{F2_Na_INAA}}{a numeric vector}
    \item{\code{F2_Na2O}}{a numeric vector}
    \item{\code{F2_Nd_INAA}}{a numeric vector}
    \item{\code{F2_Ni}}{a numeric vector}
    \item{\code{F2_Ni_INAA}}{a numeric vector}
    \item{\code{F2_NO2}}{a numeric vector}
    \item{\code{F2_NO3}}{a numeric vector}
    \item{\code{F2_P}}{a numeric vector}
    \item{\code{F2_P2O5}}{a numeric vector}
    \item{\code{F2_Pb}}{a numeric vector}
    \item{\code{F2_pH}}{a numeric vector}
    \item{\code{F2_PO4}}{a numeric vector}
    \item{\code{F2_Rb}}{a numeric vector}
    \item{\code{F2_S}}{a numeric vector}
    \item{\code{F2_Sb}}{a numeric vector}
    \item{\code{F2_Sb_INAA}}{a numeric vector}
    \item{\code{F2_Sc}}{a numeric vector}
    \item{\code{F2_Sc_INAA}}{a numeric vector}
    \item{\code{F2_Se}}{a numeric vector}
    \item{\code{F2_Se_INAA}}{a numeric vector}
    \item{\code{F2_Si}}{a numeric vector}
    \item{\code{F2_SiO2}}{a numeric vector}
    \item{\code{F2_Sm_INAA}}{a numeric vector}
    \item{\code{F2_Sn_INAA}}{a numeric vector}
    \item{\code{F2_SO4}}{a numeric vector}
    \item{\code{F2_Sr}}{a numeric vector}
    \item{\code{F2_Sr_INAA}}{a numeric vector}
    \item{\code{F2_Sum}}{a numeric vector}
    \item{\code{F2_Ta_INAA}}{a numeric vector}
    \item{\code{F2_Tb_INAA}}{a numeric vector}
    \item{\code{F2_Te}}{a numeric vector}
    \item{\code{F2_Th}}{a numeric vector}
    \item{\code{F2_Th_INAA}}{a numeric vector}
    \item{\code{F2_Ti}}{a numeric vector}
    \item{\code{F2_TiO2}}{a numeric vector}
    \item{\code{F2_U_INAA}}{a numeric vector}
    \item{\code{F2_V}}{a numeric vector}
    \item{\code{F2_W_INAA}}{a numeric vector}
    \item{\code{F2_Y}}{a numeric vector}
    \item{\code{F2_Yb_INAA}}{a numeric vector}
    \item{\code{F2_Zn}}{a numeric vector}
    \item{\code{F2_Zn_INAA}}{a numeric vector}
    \item{\code{DATE}}{a numeric vector}
    \item{\code{X.SAMP}}{a factor with levels \code{CRJHPC} \code{CRPCTF} \code{CRTF} \code{GKJHOJ} \code{GKJHTV} \code{JARR} \code{JHOJTV} \code{MÄVG} \code{MLRJARP} \code{MLRJSRR} \code{MLRMÄDR} \code{OJGKTV} \code{RPAV} \code{RPMLRJA} \code{RPVM} \code{Semenov} \code{Smirnov} \code{VGMÄ}}
    \item{\code{ELEV}}{a numeric vector}
    \item{\code{UTM}}{a numeric vector}
    \item{\code{X.COUN}}{a factor with levels \code{FIN} \code{NOR} \code{RUS}}
    \item{\code{X.ASP}}{a factor with levels \code{E} \code{FLAT} \code{N} \code{NE} \code{NW} \code{S} \code{SE} \code{SW}}
    \item{\code{X.GENLAN}}{a factor with levels \code{FLAT} \code{LOWMO} \code{PLAIN} \code{RIDGE} \code{SLOPE}}
    \item{\code{X.TOPO}}{a factor with levels \code{CONCLOW} \code{CONCMED} \code{CONVLOW} \code{CONVMED} \code{FLAT} \code{FLATLOW} \code{FLATTER} \code{LOWBRLOW} \code{LOWBRMED} \code{TER} \code{TERR} \code{TOP} \code{TOPFLAT} \code{TOPTER} \code{UPBRFLAT} \code{UPBRLOW} \code{UPBRMED} \code{UPBRTER}}
    \item{\code{X.FORDEN}}{a factor with levels \code{D} \code{MD} \code{ MD} \code{NO} \code{S}}
    \item{\code{X.TREESPE}}{a factor with levels \code{BI} \code{BI..} \code{BI.PBET.JUN} \code{BI..PI} \code{.BI.SP} \code{BI..SP} \code{BI.SP.} \code{BI.S.PJUN} \code{NO} \code{P} \code{P.} \code{P.BI} \code{P.BIJUN} \code{P.BI.S} \code{.PIBI.} \code{PI.BI} \code{PI..BI} \code{PI.BI.} \code{.PIBI.SP} \code{PI..SP} \code{PI..SPBI} \code{P.SBI} \code{P.S.BI} \code{P.SBI.JUN} \code{S.BI} \code{S.BI.JUN} \code{SP..BI} \code{SP.BI.} \code{.SPBI.PI} \code{.SPPIBI.}}
    \item{\code{TRHIGH}}{a numeric vector}
    \item{\code{RELAS}}{a numeric vector}
    \item{\code{X.BUSHDEN}}{a factor with levels \code{MD} \code{NO} \code{S}}
    \item{\code{X.BUSHSPEC}}{a factor with levels \code{} \code{BET} \code{BI} \code{..BI} \code{.BI.} \code{BI..} \code{.BI.JU} \code{BI..JU} \code{BI..PI} \code{JUN} \code{NO} \code{..RO} \code{..WI} \code{..WIBI} \code{..WIJU} \code{..WIRO} \code{..WIROJU}}
    \item{\code{X.GRVEGETATIO}}{a factor with levels \code{B..CGML} \code{B..CH} \code{B.CO.GM} \code{B.CRCHMO.LIN} \code{B.CRGRMARMO.LI} \code{B.CRMO} \code{BJUO.MO.CR} \code{B.JUOMO.LI} \code{B.LINMAR} \code{B.MO.CRMAR} \code{.BO.ML} \code{C..} \code{C..BGML} \code{C.B.GML} \code{.C.BGMLO} \code{C.B.GMLO} \code{C.B.L} \code{C.BL.GM} \code{C.BM.HGL} \code{C.BML.GO} \code{C.BO.G} \code{C.BOM.L} \code{CH.BCRLIN} \code{CH.BLIN} \code{C.L.BGM} \code{C.M.GL} \code{C..ML} \code{C.OL.M} \code{C.O.MLP} \code{CR.B.LI} \code{CR.LINMO} \code{H..BML} \code{H.L.BCM} \code{L..BMO} \code{L.BO.CM} \code{L.H.BM} \code{LIN.CR.LI} \code{M.BC.GL} \code{M..BCL} \code{M.B.CLO} \code{M.BH.CGO} \code{M.B.L} \code{M.BL.GO} \code{M.O.BCGL} \code{MO.BCR} \code{MO.BCRJUO} \code{O.B.CHMLO}}
    \item{\code{X.MOSS}}{a factor with levels \code{-9999} \code{HSDC} \code{HSDR} \code{HSSC} \code{HSSR} \code{PS} \code{PSDC} \code{PSDR} \code{PSRD} \code{PSSC}}
    \item{\code{X.TOP}}{a factor with levels \code{-9999} \code{D10} \code{D6} \code{D7} \code{M10} \code{M4} \code{M5} \code{M6} \code{M7} \code{M8}}
    \item{\code{AoMEAN}}{a numeric vector}
    \item{\code{X.AoRANGE}}{a factor with levels \code{0.1_1.0} \code{0_2} \code{0.2_2.5} \code{0.2_4.0} \code{0,5_2} \code{0,5_3} \code{0.5_4.0} \code{0.5_5.0} \code{1.0_3.0} \code{1_2} \code{1_3} \code{1_4} \code{1_5} \code{1.5_3.5} \code{1,5_5} \code{1_6} \code{2_} \code{2.0_5.0} \code{2.0_6.0} \code{2.0_7.0} \code{2_3} \code{2_4} \code{2_5} \code{2_6} \code{2_7} \code{3.0_8.0} \code{3_12} \code{3_5} \code{3_6} \code{4_12} \code{4_6} \code{4_8} \code{5_} \code{5_10} \code{.5_4} \code{-9999}}
    \item{\code{HUMNO}}{a numeric vector}
    \item{\code{HUMTHI}}{a numeric vector}
    \item{\code{X.C_PAR}}{a factor with levels \code{FLUV} \code{FLUVG} \code{TILL} \code{TILLSAP} \code{TILL&SAP}}
    \item{\code{X.C_grain}}{a numeric vector}
    \item{\code{X.COLA}}{a numeric vector}
    \item{\code{X.COLE}}{a numeric vector}
    \item{\code{LOWDE}}{a numeric vector}
    \item{\code{X.COLB}}{a numeric vector}
    \item{\code{LOWDB}}{a numeric vector}
    \item{\code{X.COLC}}{a numeric vector}
    \item{\code{TOPC}}{a numeric vector}
    \item{\code{X.WEATH}}{a factor with levels \code{DRY} \code{MIX} \code{RAIN}}
    \item{\code{TEMP}}{a numeric vector}
    \item{\code{CATLEV0}}{a numeric vector}
    \item{\code{CATLEV1}}{a numeric vector}
    \item{\code{CATLEV2}}{a numeric vector}
    \item{\code{LITO}}{a numeric vector}
    \item{\code{F1_Ag.1}}{a numeric vector}
    \item{\code{F1_Ag.2}}{a numeric vector}
    \item{\code{F2_Ag.1}}{a numeric vector}
    \item{\code{F1_Al2O3.1}}{a numeric vector}
    \item{\code{F1_Al2O3.2}}{a numeric vector}
    \item{\code{F2_Al2O3.1}}{a numeric vector}
    \item{\code{F1_Au_INAA.1}}{a numeric vector}
    \item{\code{F1_Au_INAA.2}}{a numeric vector}
    \item{\code{F2_Au_INAA.1}}{a numeric vector}
    \item{\code{F1_Ba_INAA.1}}{a numeric vector}
    \item{\code{F1_Ba_INAA.2}}{a numeric vector}
    \item{\code{F2_Ba_INAA.1}}{a numeric vector}
  }
}
\source{
Kola Project (1993-1998)
}
\references{
Reimann C, Äyräs M, Chekushin V, Bogatyrev I, Boyd R, Caritat P de, Dutter R,
Finne TE, Halleraker JH, Jæger Ø, Kashulina G, Lehto O, Niskavaara H, Pavlov V,
Räisänen ML, Strand T, Volden T. Environmental Geochemical Atlas of the Central
Barents Region. NGU-GTK-CKE Special Publication, Geological Survey of Norway,
Trondheim, Norway, 1998.
}
\author{ Peter Filzmoser <\email{P.Filzmoser@tuwien.ac.at}>
\url{http://www.statistik.tuwien.ac.at/public/filz/}}
\examples{
data(CHorFieldDUP)
str(CHorFieldDUP)
}
\keyword{datasets}
