\name{Component-class}
\docType{class}
\alias{Component-class}

\title{Class "Component" }
\description{    A Virtual base class for creating Composites }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
     \describe{
    \item{\code{LR}:}{Object of class \code{"numeric"}  }
    \item{\code{w}:}{Object of class \code{"numeric"}  }
    \item{\code{h}:}{Object of class \code{"numeric"}  }
    \item{\code{El}:}{Object of class \code{"numeric"}  }
    \item{\code{Bole}:}{Object of class \code{"numeric"}  }
  }
}
\section{Extends}{
Class \code{"\linkS4class{UComponent}"}, directly.
}
\section{Methods}{
No methods defined with class "Component" in the signature.
}
\references{ C. Reimann, P. Filzmoser, R.G. Garrett, and R. Dutter: Statistical Data Analysis
   Explained. Applied Environmental Statistics with R. John Wiley and Sons, Chichester, 2008. }
\author{ Peter Filzmoser <\email{P.Filzmoser@tuwien.ac.at}>
\url{http://cstat.tuwien.ac.at/filz/}}
\examples{
showClass("Component")
}
\keyword{classes}
