\name{StatDA-internal}
\title{Internal StatDA objects}
\alias{arbit.size}
\alias{arbit.symbol}
\alias{boxplotperc.stats}
\alias{circles}
\alias{conv}
\alias{EDAacc2.size} 
\alias{EDAacc2.symbol}
\alias{EDAacc3.size}
\alias{EDAacc3.symbol}
\alias{EDAacc.size} 
\alias{EDAacc.symbol}
\alias{EDAext.size}
\alias{EDAext.symbol}
\alias{edaplot} 
\alias{EDA.size}
\alias{EDA.symbol}
\alias{GSC.size}
\alias{GSC.symbol}
\alias{print.corsign}
\alias{seth}
\alias{setlr}
\alias{setw}
\description{Internal StatDA objects.}
\details{These are not to be called by the user.
}
\keyword{internal}
