\encoding{latin1}
\name{timetrend}
\alias{timetrend}
\docType{data}
\title{Data for computing time trends}
\description{
These are time trends from the Kola Project data.
}
\usage{data(timetrend)}
\format{
  A data frame with 96 observations on the following 47 variables.
  \describe{
    \item{\code{DD}}{a numeric vector}
    \item{\code{MM}}{a numeric vector}
    \item{\code{YY}}{a numeric vector}
    \item{\code{Year}}{a numeric vector}
    \item{\code{Catch}}{a numeric vector}
    \item{\code{X.ID}}{a numeric vector}
    \item{\code{Ag}}{a numeric vector}
    \item{\code{Al}}{a numeric vector}
    \item{\code{As}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{Ba}}{a numeric vector}
    \item{\code{Be}}{a numeric vector}
    \item{\code{Bi}}{a numeric vector}
    \item{\code{Cd}}{a numeric vector}
    \item{\code{Co}}{a numeric vector}
    \item{\code{Cr}}{a numeric vector}
    \item{\code{Cu}}{a numeric vector}
    \item{\code{Fe}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{Li}}{a numeric vector}
    \item{\code{Mn}}{a numeric vector}
    \item{\code{Mo}}{a numeric vector}
    \item{\code{Ni}}{a numeric vector}
    \item{\code{Pb}}{a numeric vector}
    \item{\code{Rb}}{a numeric vector}
    \item{\code{Sb}}{a numeric vector}
    \item{\code{Se}}{a numeric vector}
    \item{\code{Sr}}{a numeric vector}
    \item{\code{Th}}{a numeric vector}
    \item{\code{Tl}}{a numeric vector}
    \item{\code{U}}{a numeric vector}
    \item{\code{V}}{a numeric vector}
    \item{\code{Zn}}{a numeric vector}
    \item{\code{Ca}}{a numeric vector}
    \item{\code{Mg}}{a numeric vector}
    \item{\code{Na}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{S}}{a numeric vector}
    \item{\code{Si}}{a numeric vector}
    \item{\code{PO4}}{a numeric vector}
    \item{\code{Br}}{a numeric vector}
    \item{\code{Cl}}{a numeric vector}
    \item{\code{F}}{a numeric vector}
    \item{\code{NO3}}{a numeric vector}
    \item{\code{SO4}}{a numeric vector}
    \item{\code{pH}}{a numeric vector}
    \item{\code{EC}}{a numeric vector}
  }
}
\source{
Kola Project (1993-1998)
}
\references{
Reimann C, ?yr?s M, Chekushin V, Bogatyrev I, Boyd R, Caritat P de, Dutter R,
Finne TE, Halleraker JH, J?ger ?, Kashulina G, Lehto O, Niskavaara H, Pavlov V,
R?is?nen ML, Strand T, Volden T. Environmental Geochemical Atlas of the Central
Barents Region. NGU-GTK-CKE Special Publication, Geological Survey of Norway,
Trondheim, Norway, 1998.
}
\author{ Peter Filzmoser <\email{P.Filzmoser@tuwien.ac.at}>
\url{http://cstat.tuwien.ac.at/filz/}}
\examples{
data(timetrend)
str(timetrend)
}
\keyword{datasets}
