\name{UComponent-class}
\docType{class}
\alias{UComponent-class}

\title{Class "UComponent" }
\description{ To allow Component slots be also NULL }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "UComponent" in the signature.
}
\references{
C. Reimann, P. Filzmoser, R.G. Garrett, and R. Dutter: Statistical Data Analysis
   Explained. Applied Environmental Statistics with R. John Wiley and Sons Inc. To appear.}
\author{ P.Filzmoser <P.Filzmoser@tuwien.ac.at> }
\examples{
showClass("UComponent")
}
\keyword{classes}
